refactor: replace ServerAck with Ack and update batch processing in server
