+++ server/common/mbp.py
@@ -1,8 +1,32 @@
+import logging
 import struct
 
 from common.utils import Bet
-from common.safe_transport import safe_rcv
-from common.ack import SUCCESS_ID, ServerAck
+from common.safe_transport import safe_rcv, safe_send
+from common.batch import Batch
+
+# Ack
+# Paquete de reconocimiento del servidor
+# ID: 
+#   - 0 - Success
+#   - 1 - Error
+# Bytes leídos: int 
+
+SUCCESS_ID = 0
+ERROR_ID = 1
+
+class Ack:
+    def __init__(self, id: int, bets_read: int):
+        self.id = id
+        self.bets_read = bets_read
+    
+    def serialize(self) -> bytes:
+        return self.id.to_bytes(4, 'big') + self.bets_read.to_bytes(4, 'big')
+
+    def send(self, client_sock) -> None:
+        """Envía el paquete ACK al cliente."""
+        data = self.serialize()
+        safe_send(client_sock, data)
 
 def __deserialize_string(data: bytes, offset: int):
     """Lee una string: primero 2 bytes de longitud, luego contenido."""
@@ -12,7 +36,7 @@ def __deserialize_string(data: bytes, offset: int):
     offset += length
     return s, offset
 
-def deserialize_bet(data: bytes) -> Bet:
+def __deserialize_bet(data: bytes) -> Bet:
     # Leer longitud total
     total_length = struct.unpack_from(">I", data, 0)[0]
     offset = 4
@@ -33,6 +57,26 @@ def deserialize_bet(data: bytes) -> Bet:
     
     return bet, offset
 
+def __deserialize_bets_in_batch(batch: Batch) -> list[Bet]:
+    """Deserializa un Batch a partir de datos en bruto."""
+    bets = []
+    offset = 0
+
+    while offset < batch.size:
+        bet, bet_size = __deserialize_bet(batch.bytes[offset:])
+        bets.append(bet)
+        offset += bet_size
+
+    return bets
+
+def rcv_bets_in_batch(client_sock) -> list[Bet]:
+    """Recibe un Batch de Bets desde el socket del cliente."""
+    batch = Batch.recv(client_sock)
+    bets = __deserialize_bets_in_batch(batch)
+    Ack(SUCCESS_ID, len(bets)).send(client_sock)        
+    logging.info(f"action: ack_sent | result: success | ack_id: {SUCCESS_ID} | bets_read: {len(bets)}")
+    return bets        
+
 def recv_bet(client_sock):
     """Recibe un Bet serializado de forma simple."""
     # Leer longitud
@@ -43,6 +87,6 @@ def recv_bet(client_sock):
     except ConnectionError:
         raise ConnectionError("Conexión cerrada al leer longitud")
 
-    ServerAck(SUCCESS_ID, 1).send(client_sock)
+    return __deserialize_bet(payload)
 
-    return deserialize_bet(payload)
+
